refactor: export rpcmock as tgmock
