--- telegram/internal/rpcmock/doc.go
@@ -1,2 +1,0 @@
-// Package rpcmock contains RPC mocking utilities.
-package rpcmock
