+++ telegram/message/inline/inline_test.go
@@ -9,13 +9,13 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/gotd/td/bin"
-	"github.com/gotd/td/telegram/internal/rpcmock"
 	"github.com/gotd/td/tg"
 	"github.com/gotd/td/tgerr"
+	"github.com/gotd/td/tgmock"
 )
 
-func testBuilder(t *testing.T) (*ResultBuilder, *rpcmock.Mock) {
-	mock := rpcmock.NewMock(t, require.New(t))
+func testBuilder(t *testing.T) (*ResultBuilder, *tgmock.Mock) {
+	mock := tgmock.NewMock(t, require.New(t))
 	sender := New(tg.NewClient(mock), rand.Reader, 10)
 	return sender, mock
 }
