+++ tgmock/doc.go
@@ -1,0 +1,2 @@
+// Package tgmock implements mock for tg.Invoker.
+package tgmock
