+++ tgmock/handler.go
@@ -1,4 +1,4 @@
-package rpcmock
+package tgmock
 
 import "github.com/gotd/td/bin"
